+++ src/modules/sales/entities/salesman.vw.ts
@@ -1,0 +1,10 @@
+import { Field, ObjectType } from 'type-graphql';
+import { BaseEntity, Entity, PrimaryColumn } from 'typeorm';
+
+@Entity('GBR_SALESMAN_V')
+@ObjectType()
+export class Salesman extends BaseEntity {
+  @Field()
+  @PrimaryColumn({ name: 'DIVISI' })
+  divisi!: string;
+}