+++ src/modules/transportation/entities/spt-requisition.ts
@@ -13,8 +13,20 @@ export class Requisition extends BaseEntity {
   destinationId!: string;
 
   @Field()
+  @Column({ name: 'DS' })
+  ds!: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'DIVISI' })
+  divisi?: string;
+
+  @Field()
   @Column({ name: 'CUSTOMER_ID' })
   customerId!: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'VIA' })
+  via?: string;
 
   @Field()
   @Column({ name: 'REQUISITION_DATE' })
@@ -25,16 +37,20 @@ export class Requisition extends BaseEntity {
   rollQty!: number;
 
   @Field({ nullable: true })
+  @Column({ name: 'SPACE' })
+  space?: number;
+
+  @Field({ nullable: true })
   @Column({ name: 'METER' })
   meter?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'WEIGHT' })
-  weight!: number;
+  weight?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'VOLUME' })
-  volume!: number;
+  volume?: number;
 
   @Field()
   @Column({ name: 'CONTRACT' })