+++ src/modules/transportation/entities/spt-unassign-view.ts
@@ -21,12 +21,24 @@ export class UnassignView extends BaseEntity {
   destinationName!: string;
 
   @Field()
+  @Column({ name: 'DS' })
+  ds!: string;
+
+  @Field()
   @Column({ name: 'CUSTOMER_ID' })
   customerId!: string;
 
   @Field()
   @Column({ name: 'CUSTOMER_NAME' })
   customerName!: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'VIA' })
+  via?: string;
+
+  @Field()
+  @Column({ name: 'DIVISI' })
+  divisi!: string;
 
   @Field()
   @Column({ name: 'REQUISITION_DATE' })
@@ -36,15 +48,23 @@ export class UnassignView extends BaseEntity {
   @Column({ name: 'ROLL_QTY' })
   rollQty!: number;
 
-  @Field()
+  @Field({ nullable: true })
+  @Column({ name: 'SPACE' })
+  space?: number;
+
+  @Field({ nullable: true })
   @Column({ name: 'METER' })
-  meter!: number;
+  meter?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'WEIGHT' })
-  weight!: number;
+  weight?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'VOLUME' })
-  volume!: number;
+  volume?: number;
+
+  @Field({ nullable: true })
+  @Column({ name: 'NOTES' })
+  notes?: string;
 }