+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="1.2.2"></a>
+# [1.2.2](https://gitlab.com/atjdev/ezio-api/-/merge_requests/49) (2021-10-26)
+
+### Features
+
+* [**T001**] SPT: Add "via" column for requisition resolver and related views
+
+<a name="1.2.1"></a>
+# [1.2.1](https://gitlab.com/atjdev/ezio-api/-/merge_requests/47) (2021-10-23)
+
+### Features
+
+* [**T001**] SPT: add ds, divisi, and space column for requisition
+* [**T001**] SPT: add nopol langsir for surat ijin keluar
+
 <a name="1.2.0"></a>
 # [1.2.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/45) (2021-10-12)
 
@@ -14,6 +29,7 @@
 * [**T001**] SPT: Add expedition name and vehicle name column for assign requisition view
 
 <a name="1.1.4"></a>
+
 # [1.1.4](https://gitlab.com/atjdev/ezio-api/-/merge_requests/41) (2021-10-07)
 
 ### BREAKING CHANGES